+++ InventoryControl/Validators/Api/CategoryDtoValidator.cs
@@ -11,5 +11,6 @@ public class CategoryDtoValidator : AbstractValidator<CategoryDto>
     {
         RuleFor(x => x.Name).NotEmpty().MaximumLength(100);
         RuleFor(x => x.Description).MaximumLength(500).When(x => x.Description is not null);
+        RuleFor(x => x.ParentId).GreaterThan(0).When(x => x.ParentId is not null);
     }
 }